+++ CHANGELOG.md
@@ -184,6 +184,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   Please [see the documentation](/docs/config.md#json-schema-support) for more
   on this.
 
+* The new `jj sign` command allows signing commits.
+
+* The new `jj unsign` command allows unsigning commits.
+
 ### Fixed bugs
 
 * `jj git fetch` with multiple remotes will now fetch from all remotes before